数据库编程式和声明式事务实践

+++ spring-boot-mybatis-plus/src/main/java/org/chy/carorder/service/CarOrderServices.java
@@ -26,5 +26,5 @@ public interface CarOrderServices {
 
     boolean add(CarOrderAddReqDto reqDto);
 
-    Long submit(CarOrderSubmitReqDto reqDto);
+    Integer submit(CarOrderSubmitReqDto reqDto);
 }
